--- client/src/common/oldButton.jsx
@@ -1,17 +1,0 @@
-
-const name = ref('Vue.js')
-import '../../styles/Button.scss';
-
-function ApplyButton() {
-    return (
-        <div class="button-container-1">
-            <span class="mas">Apply!</span>
-            <button type="button" name="Hover">Apply!</button>
-        </div> 
-    )
-}
- 
-export {ApplyButton};
-
-
-
