+++ client/src/styles/Button.scss
@@ -35,6 +35,7 @@
         border-radius: 8px;
     
         button {
+                color: red;
                 width: 101%;
                 height: 100%;  
                 font-size: 30px;
